feat: payments, maintenance blocks, hold expiry, and role-based access control
Add four related capabilities on top of the count-based inventory engine.

Payment hold lifecycle + optimistic locking:
- New bookings start as PENDING holds with a hold_expires_at window instead of
  going straight to CONFIRMED; add EXPIRED status for lapsed holds.
- HoldExpirySweeper (@Scheduled) releases inventory for holds whose window has
  passed; expireHold runs per-reservation so one failure doesn't roll back the
  sweep. A @Version column guards against a payment and the sweep racing on the
  same PENDING row.
- Cancellation now accepts PENDING holds as well as CONFIRMED bookings.

Payments:
- PaymentController/Service with balance-due tracking; a full payment moves the
  reservation to CONFIRMED and clears the hold, partial payment leaves it PENDING.
- Payment sums use BigDecimal end-to-end so money never round-trips through double.

Maintenance blocks:
- MaintenanceController/Service to take a room out of service; decrements sellable
  capacity per night and rejects a block when a night is already fully sold.
  Overlapping blocks on the same room are rejected. Lifting a block returns the
  capacity via InventoryService.increment/decrementCapacity.

Role-based access control:
- Two-role model (ROLE_USER | ROLE_ADMIN) backed by a persisted role column;
  authorities derive from it. Self-registration is pinned to ROLE_USER; the only
  admin is seeded via data.sql.
- Admin-only write endpoints for the hotel catalogue (POST/PUT/DELETE on hotels
  and nested room-types, guarded against cross-hotel mutation).
- Layered enforcement: SecurityConfig gates catalogue writes, maintenance, and
  reservation check-in by role, with @PreAuthorize repeated at each method.
  AccessDeniedException now maps to a clean 403.

Tests: role assertions, Guest.getAuthorities(), hotel/room-type write and
cross-hotel guard coverage, hold expiry, payments, and maintenance capacity.
Full suite green (88 tests).

+++ src/main/java/com/Abdelwahab/RoomBooking/repository/PaymentRepository.java
@@ -1,5 +1,6 @@
 package com.Abdelwahab.RoomBooking.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -14,7 +15,9 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
 
     List<Payment> findByReservationId(Long reservationId);
 
-    // Sum all successful payments for a reservation (used to calculate balance due at checkout)
+    // Sum all successful payments for a reservation (used to calculate balance due at checkout).
+    // Returns BigDecimal to stay in exact decimal arithmetic — money must never round-trip through double.
+    // COALESCE(..., 0) turns the no-rows NULL into a zero total.
     @Query("SELECT COALESCE(SUM(p.amount), 0) FROM Payment p WHERE p.reservation.id = :reservationId AND p.status = 'SUCCESS'")
-    double sumSuccessfulPaymentsByReservationId(@Param("reservationId") Long reservationId);
+    BigDecimal sumSuccessfulPaymentsByReservationId(@Param("reservationId") Long reservationId);
 }
